Develop room kmp

+++ feature/edit-note/src/commonMain/kotlin/com/khomichenko/edit_note/component/EditNoteComponentImpl.kt
@@ -11,7 +11,7 @@ import org.koin.core.component.get
 
 internal class EditNoteComponentImpl(
     componentContext: ComponentContext,
-    private val idNote: Int,
+    private val idNote: Long,
     private val closeSlotComponent:() -> Unit
 ) : EditNoteComponent, KoinComponent, ComponentContext by componentContext {
