+++ feature/notes/src/commonMain/kotlin/com/khomichenko/notes/component/NotesComponent.kt
@@ -12,7 +12,7 @@ interface NotesComponent {
     val state: Flow<Model>
 
     fun openBottomSheetComponent()
-    fun openEditSlotComponent(id: Int)
+    fun openEditSlotComponent(id: Long)
 }
